Add : max_samples as parameter to RandomForest

+++ src/split_strategy/types/split_hyper.h
@@ -16,6 +16,6 @@ struct HyperParam{
     std::optional<int> mtry=std::nullopt;
     std::optional<int> n_estimators=std::nullopt;
     std::optional<int> max_depth=std::nullopt;
-    std::optional<int> max_samples=std::nullopt;
+    std::optional<float> max_samples=std::nullopt;
     
 };
